chore(helpers): refactor min max string generation
