+++ src/stores/datarequests.ts
@@ -196,7 +196,7 @@ export const getParsedStringItem = (
             return valueItem.name;
         }
 
-        return getMinMax(valueItem.value);
+        return valueItem.name;
     });
 
     return `${name}: ${values.join(", ")}`;
@@ -238,23 +238,4 @@ const getParsedAggregatedStringValues = (
             `\n\t\t${translate("query_item_multi_row_header")}\n`,
         );
     return parsedAggregatedGroupsString;
-};
-
-/**
- * Format a min/max object as a string.
- * @param {{min?: string | number, max?: string | number}} param0
- *   min: The minimum value (string or number, optional).
- *   max: The maximum value (string or number, optional).
- * @returns {string} A string representation of the min/max range.
- */
-export const getMinMax = ({
-    min,
-    max,
-}: {
-    min?: string | number;
-    max?: string | number;
-}): string => {
-    if (min && max) return `${min} - ${max}`;
-    if (!min) return `≤ ${max}`;
-    return `≥ ${min}`;
 };
